requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-debug-toolbar==3.1.1
django-extensions==3.1.0
django-markdownx==3.0.1
django-taggit==1.3.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
Markdown==3.3.3
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytils==0.3
pytz==2019.3
requests==2.31.0
requests-oauthlib==1.3.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.3.0
urllib3==1.25.6
wcwidth==0.1.8
webencodings==0.5.1
zipp==2.2.0
gunicorn
sentry-sdk
